+++ packages/gasket-cli/src/config/default-plugins.js
@@ -1,5 +1,6 @@
 module.exports = [
   require('@gasket/command-plugin'),
   require('@gasket/lifecycle-plugin'),
-  require('@gasket/metadata-plugin')
+  require('@gasket/metadata-plugin'),
+  require('@gasket/start-plugin')
 ];